+++ sycl/include/sycl/ext/oneapi/annotated_arg/properties.hpp
@@ -154,7 +154,6 @@ template <typename T, typename PropertyListT>
 struct is_property_key_of<wait_request_key, annotated_arg<T, PropertyListT>>
     : std::true_type {};
 
-
 template <typename T, typename PropertyListT>
 struct is_property_key_of<register_map_key, annotated_ptr<T, PropertyListT>>
     : std::true_type {};
@@ -194,8 +193,6 @@ struct is_property_key_of<maxburst_key, annotated_ptr<T, PropertyListT>>
 template <typename T, typename PropertyListT>
 struct is_property_key_of<wait_request_key, annotated_ptr<T, PropertyListT>>
     : std::true_type {};
-
-
 
 namespace detail {
 template <> struct PropertyToKind<register_map_key> {
